backend/requirements.txt
psycopg2
Flask
SQLAlchemy
python-dotenv
flask-migrate
#pip install -r requirements.txt 
#pip freeze